fix(pages+visitor): pluggable widget TabIndex int64→int32 + log with() nil panic
Bug 4 — CustomWidget TabIndex stored as int64 causes NullReferenceException
in Studio Pro when opening imported pages:
  - widget_builder.go FinalizeGen and datagrid_builder.go buildDataGrid2WidgetDoc
    both set TabIndex = int64(0). Mendix expects int32 for this field.
  - When Studio Pro tries to set the int64 value on an int32 property,
    SetPropertyValue throws an exception. The exception handler calls
    Description.Type(widget) → GetCustomDescription(), but at that point
    the Type field (which comes after TabIndex in the BSON doc) has not
    been processed yet, so WidgetType is null → NullReferenceException.
  Fix: int64(0) → int32(0) in both locations.

Bug 5 — nil pointer panic in buildTemplateParams with empty `with ()` clause:
  - ANTLR error recovery for `with ()` (grammar error — requires at least
    one {N} = expr) may insert a synthetic TemplateParamContext without a
    NUMBER_LITERAL token.
  - paramCtx.NUMBER_LITERAL().GetText() panicked on the nil terminal.
  Fix: guard with numLit := paramCtx.NUMBER_LITERAL(); if numLit == nil { continue }.

+++ mdl/backend/mpr/datagrid_builder.go
@@ -74,7 +74,7 @@ func (b *MprBackend) buildDataGrid2WidgetDoc(id model.ID, name string, spec back
 		{Key: "LabelTemplate", Value: nil},
 		{Key: "Name", Value: name},
 		{Key: "Object", Value: updatedObject},
-		{Key: "TabIndex", Value: int64(0)},
+		{Key: "TabIndex", Value: int32(0)},
 		{Key: "Type", Value: embeddedType},
 	}
 	return doc, nil
